Move copy/paste for NamedElement to UML package
